// chicagovisa-dashboard/src/types/service-keys.ts
export type ServiceKeys =
  | "stats"
  | "cases"
  | "forms"
  | "promoCodes"
  | "serviceLevel"
  | "serviceTypes"
  | "statuses"
  | "reports"
  | "faq"
  | "loa"
  | "files"
  | "roles"
  | "users"
  | "configuration"
  | "countries"
  | "countryAccess"
  | "logs"
  | "shippingLocations"
  | "additionalServices"
  | "fedexPackages"
  | "manifest"
  | "queries";